Move sanitization message to dataset display
